+++ src/ts/components/seekbar.ts
@@ -1,10 +1,10 @@
-import {Component, ComponentConfig} from './component';
-import {DOM} from '../dom';
-import {Event, EventDispatcher, NoArgs} from '../eventdispatcher';
-import {SeekBarLabel} from './seekbarlabel';
-import {UIInstanceManager, SeekPreviewArgs} from '../uimanager';
-import {Timeout} from '../timeout';
-import {PlayerUtils} from '../playerutils';
+import { Component, ComponentConfig } from './component';
+import { DOM } from '../dom';
+import { Event, EventDispatcher, NoArgs } from '../eventdispatcher';
+import { SeekBarLabel } from './seekbarlabel';
+import { UIInstanceManager, SeekPreviewArgs } from '../uimanager';
+import { Timeout } from '../timeout';
+import { PlayerUtils } from '../playerutils';
 import TimeShiftAvailabilityChangedArgs = PlayerUtils.TimeShiftAvailabilityChangedArgs;
 import LiveStreamDetectorEventArgs = PlayerUtils.LiveStreamDetectorEventArgs;
 import { TimelineMarker } from '../uiconfig';
@@ -13,6 +13,7 @@ import { StringUtils } from '../stringutils';
 import { SeekBarType, SeekBarController } from './seekbarcontroller';
 import { i18n } from '../localization/i18n';
 import { BrowserUtils } from '../browserutils';
+import { TimelineMarkersHandler } from './timelinemarkershandler';
 
 /**
  * Configuration interface for the {@link SeekBar} component.
@@ -86,11 +87,11 @@ export class SeekBar extends Component<SeekBarConfig> {
   private seekBarBufferPosition: DOM;
   private seekBarSeekPosition: DOM;
   private seekBarBackdrop: DOM;
-  private seekBarMarkersContainer: DOM;
 
   private label: SeekBarLabel;
 
-  private timelineMarkers: SeekBarMarker[];
+  private seekBarMarkersContainer: DOM;
+  private timelineMarkersHandler: TimelineMarkersHandler;
 
   private player: PlayerAPI;
 
@@ -142,7 +143,6 @@ export class SeekBar extends Component<SeekBarConfig> {
     }, this.config);
 
     this.label = this.config.label;
-    this.timelineMarkers = [];
   }
 
   initialize(): void {
@@ -288,7 +288,7 @@ export class SeekBar extends Component<SeekBarConfig> {
       this.setSeeking(false);
     };
 
-    let restorePlayingState = function() {
+    let restorePlayingState = function () {
       // Continue playback after seek if player was playing when seek started
       if (isPlaying) {
         // use the same issuer here as in the pause on seek
@@ -369,10 +369,9 @@ export class SeekBar extends Component<SeekBarConfig> {
     });
     let timeShiftDetector = new PlayerUtils.TimeShiftAvailabilityDetector(player);
     timeShiftDetector.onTimeShiftAvailabilityChanged.subscribe((sender, args: TimeShiftAvailabilityChangedArgs) => {
-        hasTimeShift = args.timeShiftAvailable;
-        switchVisibility(isLive, hasTimeShift);
-      },
-    );
+      hasTimeShift = args.timeShiftAvailable;
+      switchVisibility(isLive, hasTimeShift);
+    });
     // Initial detection
     liveStreamDetector.detect();
     timeShiftDetector.detect();
@@ -408,7 +407,18 @@ export class SeekBar extends Component<SeekBarConfig> {
     if (this.config.smoothPlaybackPositionUpdateIntervalMs !== SeekBar.SMOOTH_PLAYBACK_POSITION_UPDATE_DISABLED) {
       this.configureSmoothPlaybackPositionUpdater(player, uimanager);
     }
-    this.configureMarkers(player, uimanager);
+
+    // Initialize markers
+    this.initializeTimelineMarkers(player, uimanager);
+  }
+
+  private initializeTimelineMarkers(player: PlayerAPI, uimanager: UIInstanceManager): void {
+    const timelineMarkerConfig = {
+      cssPrefix: this.config.cssPrefix,
+      snappingRange: this.config.snappingRange,
+    };
+    this.timelineMarkersHandler = new TimelineMarkersHandler(timelineMarkerConfig, () => this.seekBar.width(), this.seekBarMarkersContainer);
+    this.timelineMarkersHandler.initialize(player, uimanager);
   }
 
   private seekWhileScrubbing = (sender: SeekBar, args: SeekPreviewEventArgs) => {
@@ -532,44 +542,6 @@ export class SeekBar extends Component<SeekBarConfig> {
     return PlayerUtils.getCurrentTimeRelativeToSeekableRange(this.player);
   }
 
-  private configureMarkers(player: PlayerAPI, uimanager: UIInstanceManager): void {
-    let clearMarkers = () => {
-      this.timelineMarkers = [];
-      this.updateMarkers();
-    };
-
-    let setupMarkers = () => {
-      clearMarkers();
-
-      const duration = player.getDuration();
-
-      if (duration === Infinity) {
-        // Don't generate timeline markers if we don't yet have a duration
-        // The duration check is for buggy platforms where the duration is not available instantly (Chrome on Android 4.3)
-        return;
-      }
-
-      for (let marker of uimanager.getConfig().metadata.markers) {
-        const markerPosition = 100 / duration * marker.time; // convert absolute time to percentage
-        const markerDuration = 100 / duration * marker.duration;
-        this.timelineMarkers.push({ marker, position: markerPosition, duration: markerDuration });
-      }
-
-      // Populate the timeline with the markers
-      this.updateMarkers();
-    };
-
-    // Remove markers when unloaded
-    player.on(player.exports.PlayerEvent.SourceUnloaded, clearMarkers);
-    // Update markers when the size of the seekbar changes
-    player.on(player.exports.PlayerEvent.PlayerResized, () => this.updateMarkers());
-    uimanager.getConfig().events.onUpdated.subscribe(setupMarkers);
-    uimanager.onRelease.subscribe(() => uimanager.getConfig().events.onUpdated.unsubscribe(setupMarkers));
-
-    // Init markers at startup
-    setupMarkers();
-  }
-
   release(): void {
     super.release();
 
@@ -664,7 +636,7 @@ export class SeekBar extends Component<SeekBarConfig> {
       new DOM(document).off('touchend mouseup', mouseTouchUpHandler);
 
       let targetPercentage = 100 * this.getOffset(e);
-      let snappedChapter = this.getMarkerAtPosition(targetPercentage);
+      let snappedChapter = this.timelineMarkersHandler && this.timelineMarkersHandler.getMarkerAtPosition(targetPercentage);
 
       this.setSeeking(false);
       seeking = false;
@@ -734,55 +706,6 @@ export class SeekBar extends Component<SeekBarConfig> {
     }
 
     return seekBarContainer;
-  }
-
-  protected updateMarkers(): void {
-    this.seekBarMarkersContainer.empty();
-
-    const seekBarWidthPx = this.seekBar.width();
-
-    for (let marker of this.timelineMarkers) {
-      const markerClasses = ['seekbar-marker'].concat(marker.marker.cssClasses || [])
-        .map(cssClass => this.prefixCss(cssClass));
-
-      const cssProperties: {[propertyName: string]: string} = {
-        'width': marker.position + '%',
-      };
-
-      if (marker.duration > 0) {
-        const markerWidthPx = Math.round(seekBarWidthPx / 100 * marker.duration);
-        cssProperties['border-right-width'] = markerWidthPx + 'px';
-        cssProperties['margin-left'] = '0';
-      }
-
-      this.seekBarMarkersContainer.append(new DOM('div', {
-        'class': markerClasses.join(' '),
-        'data-marker-time': String(marker.marker.time),
-        'data-marker-title': String(marker.marker.title),
-      }).css(cssProperties));
-    }
-  }
-
-  protected getMarkerAtPosition(percentage: number): SeekBarMarker | null {
-    const snappingRange = this.config.snappingRange;
-
-    if (this.timelineMarkers.length > 0) {
-      for (let marker of this.timelineMarkers) {
-        // Handle interval markers
-        if (marker.duration > 0
-          && percentage >= marker.position - snappingRange
-          && percentage <= marker.position + marker.duration + snappingRange) {
-          return marker;
-        }
-        // Handle position markers
-        else if (percentage >= marker.position - snappingRange
-          && percentage <= marker.position + snappingRange) {
-          return marker;
-        }
-      }
-    }
-
-    return null;
   }
 
   /**
@@ -999,7 +922,7 @@ export class SeekBar extends Component<SeekBarConfig> {
   }
 
   protected onSeekPreviewEvent(percentage: number, scrubbing: boolean) {
-    let snappedMarker = this.getMarkerAtPosition(percentage);
+    let snappedMarker = this.timelineMarkersHandler && this.timelineMarkersHandler.getMarkerAtPosition(percentage);
 
     let seekPositionPercentage = percentage;